添加|更新touchgfx示例和库 (#34)
* 添加touchgfx-factory代码；规范lvgl触摸

+++ libraries/touchgfx_lib/Middlewares/ST/touchgfx/framework/tools/textconvert/lib/file_io.rb
@@ -1,31 +1,29 @@
-##############################################################################
-# This file is part of the TouchGFX 4.15.0 distribution.
+# Copyright (c) 2018(-2024) STMicroelectronics.
+# All rights reserved.
 #
-# <h2><center>&copy; Copyright (c) 2020 STMicroelectronics.
-# All rights reserved.</center></h2>
+# This file is part of the TouchGFX 4.24.2 distribution.
 #
-# This software component is licensed by ST under Ultimate Liberty license
-# SLA0044, the "License"; You may not use this file except in compliance with
-# the License. You may obtain a copy of the License at:
-#                             www.st.com/SLA0044
+# This software is licensed under terms that can be found in the LICENSE file in
+# the root directory of this software component.
+# If no LICENSE file comes with this software, it is provided AS-IS.
 #
-##############################################################################
-
+###############################################################################/
 require 'rubygems'
 require 'erb'
 require 'fileutils'
 require 'pathname'
 
-
 class FileIO
   def self.write_file(file_name, contents)
-    write_file_silent(file_name, contents)
     callingPath = Pathname.new($calling_path)
     filePath = Pathname.new(file_name)
     puts "Generating #{filePath.relative_path_from(callingPath)}"
+    write_file_silent(file_name, contents)
   end
   def self.write_file_silent(file_name, contents)
     FileUtils.mkdir_p(File.dirname(file_name))
-    File.open(file_name, 'w') { |f| f.write(contents) }
+    unless File.exist?(file_name) && contents == File.open(file_name, 'r') { |f| f.read() }
+      File.open(file_name, 'w') { |f| f.write(contents) }
+    end
   end
 end
